MVC 패턴구현

+++ src/model/Operator.java
@@ -1,34 +1,32 @@
 package model;
 
-import controller.Num;
-
 import java.util.function.BiFunction;
 import java.util.Arrays;
 
 public enum Operator {
 
-    PLUS("+", (num1, num2) -> num1 + num2),
-    MINUS("-", (num1, num2) -> num1 - num2),
-    MULTIPLY("*", (num1, num2) -> num1 * num2),
-    DIVIDE("/", (num1, num2) -> num1 / num2);
+  PLUS("+", (num1, num2) -> num1 + num2),
+  MINUS("-", (num1, num2) -> num1 - num2),
+  MULTIPLY("*", (num1, num2) -> num1 * num2),
+  DIVIDE("/", (num1, num2) -> num1 / num2);
 
-    private final String operator;
-    private final BiFunction<Integer, Integer, Integer> expression;
+  private final String operator;
+  private final BiFunction<Double, Double, Double> expression;
 
-    Operator(String operator, BiFunction<Integer, Integer, Integer> expression) {
-        this.operator = operator;
-        this.expression = expression;
-    }
+  Operator(String operator, BiFunction<Double, Double, Double> expression) {
+    this.operator = operator;
+    this.expression = expression;
+  }
 
-    public static Integer calculate(Integer num1, String operator, Integer num2) {
-        return getOperator(operator)
-                .expression
-                .apply(num1, num2);
-    }
+  public static Double calculate(Integer num1, String operator, Integer num2) {
+    return getOperator(operator)
+        .expression
+        .apply((double) num1, (double) num2);
+  }
 
-    private static Operator getOperator(String operator) {
-        return Arrays.stream(values())
-                .filter(o -> o.operator.equals(operator))
-                .findFirst().orElseThrow(() -> new IllegalArgumentException("올바른 연산자가 아닙니다."));
-    }
+  private static Operator getOperator(String operator) {
+    return Arrays.stream(values())
+        .filter(o -> o.operator.equals(operator))
+        .findFirst().orElseThrow(() -> new IllegalArgumentException("올바른 연산자가 아닙니다."));
+  }
 }
